fixed clicking of objects to Spatial and execution of event depends on instance of RayCast3DNode; added MapSquare class to Map for the graphical representation of a grid-square; fixed isocamera being active in debug mode; some minor fixes;
git-svn-id: https://subversion.assembla.com/svn/mazetd/trunk@11 65d133f4-921b-4a0d-80d6-c7a6dbb31848

+++ mazetd/src/events/TimerEventListener.java
@@ -44,5 +44,5 @@ public interface TimerEventListener extends EventListener {
          * 
          * @return the period
          */
-        public long getPeriod();
+        public float getPeriod();
 }
